🍗 :: [#158] 타이머 모델
자료형 변경

+++ Projects/Features/HomeFeature/Sources/TimerScene/Supporter/TimerModel.swift
@@ -3,7 +3,7 @@ import UIKit
 import DSKit
 
 public struct Time {
-    var time: Double
+    var time: Int
 }
 
 public enum TimerModel {
